notes - refactoring code (imports - exports) - index.ts
