+++ src/enums/enum.ts
@@ -1,0 +1,5 @@
+export enum WidgetTitles {
+    WeatherWidget = 'weather-widget',
+    NewsWidget = 'news-widget',
+    NotesWidget = 'notes-widget',
+}
