+++ src/enums/index.ts
@@ -1,0 +1,1 @@
+export {WidgetTitles} from './enum';
