+++ src/typings/index.ts
@@ -1,0 +1,7 @@
+import {WidgetTitles} from "src/enums";
+
+export type PropsType = {
+    app?: WidgetTitles;
+    city?: string;
+    country?: string;
+};
