+++ src/utils/getLowerCaseString.ts
@@ -1,3 +1,7 @@
-export const getLowerCaseString = (value: string | undefined) => {
+export const getLowerCaseString = (value?: string) => {
+    if (!value) {
+        return;
+    }
+
     return value?.toLowerCase();
 };
