+++ src/widgets/index.ts
@@ -1,0 +1,3 @@
+export {News} from './news/News';
+export {Notes} from './notes/Notes';
+export {Weather} from './weather/Weather';
