+++ src/widgets/notes/components/accordionWithNotes/AccordionWithNotes.module.css
