+++ src/widgets/notes/components/accordionWithNotes/accordionItem/AccordionItem.module.css
