+++ src/widgets/notes/components/accordionWithNotes/accordionItem/index.ts
@@ -1,0 +1,1 @@
+export {AccordionItem} from './AccordionItem';
