+++ src/widgets/notes/components/accordionWithNotes/accordionItem/typings/index.ts
@@ -1,0 +1,10 @@
+import {NotesType} from 'src/widgets/notes/typings';
+
+export type PropsType = {
+    key: number;
+    city?: string;
+    country: string;
+    findedCity: string;
+    notes: NotesType[];
+    removeNote: (title?: string) => void;
+};
