+++ src/widgets/notes/components/accordionWithNotes/index.ts
