+++ src/widgets/notes/components/index.ts
@@ -1,2 +1,1 @@
-export {AccordionWithNotes} from './accordionWithButtonForm';
-// export {FormWithNotes} from './formWithNotes';
+export {AccordionWithNotes} from './accordionWithNotes';
