+++ src/widgets/notes/service/getNotes.ts
@@ -1,8 +1,8 @@
 import {NotesType} from '../typings';
-import {getParseLocalStorageData} from '../../../services/localStorage';
+import {getParseLocalStorageData} from 'src/services/localStorage';
 
 export const getNotes = (localStorageName: string, newNote?: NotesType,
-                         city?: string | undefined, setNotes?: Function) => {
+                         city?: string, setNotes?: Function) => {
     const getLocalStorageData = getParseLocalStorageData(localStorageName);
 
     const findedNotes = getLocalStorageData
